requirements.txt
python-telegram-bot[ext]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
python-dotenv
matplotlib # Add matplotlib for charting
